+++ src/app/api/user/[userId]/route.ts
@@ -1,5 +1,5 @@
 import { ObjectId, Document } from "mongodb";
-import { connectToDatabase } from "../../../../../utils/mongodb";
+import { connectToDatabase } from "@/utils/mongodb";
 import { NextRequest } from "next/server";
 
 export async function GET(
