+++ src/app/api/user/search/route.ts
@@ -1,4 +1,4 @@
-import { connectToDatabase } from "../../../../../utils/mongodb";
+import { connectToDatabase } from "@/utils/mongodb";
 import { NextRequest } from "next/server";
 
 export async function GET(req: NextRequest) {
