+++ src/app/post/new/page.tsx
@@ -7,8 +7,8 @@ import ProfileAndName from "@/components/ProfileAndName/ProfileAndName.component
 import Textarea from "@/components/Textarea/Textarea.component";
 import React from "react";
 import PostDragAndDrop from "@/components/DragAndDrop/PostDragAndDrop/PostDragAndDrop.component";
-import { ImageUpload } from "../../../../utils/upload";
-import { ICommonResponse } from "../../../../typescript/common/response.interface";
+import { ImageUpload } from "@/utils/upload";
+import { ICommonResponse } from "@/typescript/common/response.interface";
 
 export default function NewPost() {
   /** router */
