+++ src/typescript/common/response.interface.ts
