+++ src/typescript/user.interface.ts
